Using std::stoll instead std::stoi for parse game id. Rename logs table, fix payout.

+++ include/dicegame.hpp
@@ -148,7 +148,7 @@ class[[eosio::contract]] dicegame : public contract
         *player_seed = part;
         part = data.substr(++pos);
         eosio_assert(!part.empty(), "No gameID");
-        *game_id = stoi(part);
+        *game_id = stoll(part);
     }
 };
